Create stock.py
Diversification is a key step for constructing portfolios but true diversification is not possible as investing in each company requires a lot of capital. Apart from this, not all companies are worth investing in and that may be because they are highly illiquid.

So now what you want to do is come up with an effective strategy where you pick only those stocks that behave differently. This will help in mitigating the risk and one way of doing it is to pick stocks from different sectors but a more data-driven solution can be to apply K-Means clustering algorithm on stock data to identify different clusters of stocks.

[Reference: https://www.pythonforfinance.net/2018/02/08/stock-clusters-using-k-means-algorithm-in-python/]

In the following post, I will give an example of applying the K-Means algorithm using Python.